third_party/WebKit/LayoutTests/http/tests/media/video-load-with-userpass.html

<html>
    <head>
        <script src=../../media-resources/video-test.js></script>
        <script src=../../media-resources/media-file.js></script>
        <script>
            function loadMediaFrame()
            {
                findMediaElement();

                var movie = findMediaFile('video', 'test');
                var type = mimeTypeForExtension(movie.split('.').pop());
                var frame = document.createElement('iframe');
                frame.width = 0;
                frame.height = 0;
                frame.addEventListener('load', function () {
                        source = document.getElementById('source');
                        source.src = 'http://user:pass@127.0.0.1:8080/media/resources/video-check-userpass.php?name=' + movie + '&type=' + type;
                        source.type = type;

                        waitForEventAndFail('error');
                        waitForEventAndEnd('canplay');
                        video.load();
                });

                frame.src = "data:text/html,<b>test</b>";
                document.body.appendChild(frame);
            }
        </script>
    </head>

    <body onload="loadMediaFrame()">
        <video id="video">
            <source id="source">
        </video>
        <br>
        Tests that the media player does not include authorization information.
    </body>
</html>
