third_party/WebKit/LayoutTests/http/tests/media/video-load-with-userpass-expected.txt

 
Tests that the media player does not include authorization information.  
EVENT(canplay)
END OF TEST
